pyproject.toml
[project]
name = "paperly"
version = "0.1.0"
description = "AI-powered document quality improvement for Paperless-NGX"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "anthropic>=0.37",
    "jinja2>=3.1",
    "python-dotenv>=1.0",
    "python-multipart>=0.0.9",
]

[project.scripts]
paperly = "paperly.app:main"
paperly-cleanup = "paperly.cleanup:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/paperly"]
